requirements.txt
torch
torchvision
einops
timm
tqdm
matplotlib
numpy
kagglehub
pillow
